requirements.txt
# Environment management
python-dotenv==1.0.0

# Database handling
psycopg2-binary==2.9.7

# API and HTTP requests
requests==2.31.0

# Code linting and formatting
black==23.9.1
pylint==2.17.5

# JSON parsing and handling (standard library, but some packages might be useful if needed)
simplejson==3.19.2

# Optional (if you are handling JSON serialization in other areas)
jsonschema==4.19.0

# If you are using timezone-aware date formatting with `pytz`
pytz==2023.3
